Fix bug in cargo make build-mobile and CI. (#181)
`cargo ndk`  tried to build all crates in the workspace
in cargo-make and all crates not explicitly excluded in
the CI, which lead to an error. 

`cargo lipo` had a similar problem but would only consider
crates which produce `staticlibs` and as such it didn't
cause any problems, as currently only the intended
crate produces such outputs.

We updated all usages of both commands to only
look at `xayn-ai-ffi-c` (and its dependencies).