Used RUSTC_BOOTSTRAP as a simple way to get a nightly matching the stable we use.
We are fully aware that RUSTC_BOOTSTRAP is not supposed to be used and does disable/
brake the forward guarantees stable rust gives. So we treat it as a nightly rust
release.  In the end it's the simplest way to get a nightly matching the current
stable release, which is more or less what we use it for.