TY-2128 add web worker asset type (#272)
Tickets:
- [TY-2128]
- [TY-2127]

Summary:

- moved `RerankMode` and `selectThreadPoolSize` into a separate file. That change was necessary as `RerankMode` or `selectThreadPoolSize` would pull in the code of `SetupData` when compiling the worker. `SetupData` uses the generated `AssetType` which cannot exist in that moment as we first need to compile the worker because it is an asset as well. 
- added a new `webWorkerScript` asset type that points to the url of the web worker script
- added a release ci job for compiling the dart worker
- updated the changelog

Note why we move the worker into the respective WASM artefact folders:

My first idea that I had was simply adding a separate folder for the web worker and upload it on S3:
```
<base>/wasm_<feature>_<hash>/genesis.js
<base>/web_worker_<hash>/worker.js
```

However, this didn't work as the path of `genesis.js`, that is imported via `importScripts`, needs to be relative to the path of `worker.js`.

My second idea was then to put the WASM artefacts into the web worker folder:
```
<base>/web_worker_<hash>/worker.js
<base>/web_worker_<hash>/wasm_<feature>_<hash>/genesis.js
```

Here, the import of `genesis.js` works but the import of the snippet (in case of the multithread version) does not.
When webpack bundles the multithread version it creates 3 files:

```
genesis.js
<1>_genesis.js <- snippet
<2>_genesis.js <- relative import of genesis.js
```

Because webpack does not know anything about our later folder structure it simply uses `self.location + ''` as a base path. `self.location + ''` resolves to the path of `worker.js`. Once `genesis.js` tries to load `<1>_genesis.js` it uses the URL `<base>/web_worker_<hash>/<1>_genesis.js` instead of `<base>/web_worker_<hash>/wasm_<feature>_<hash>/<1>_genesis.js`. Webpack offers an option [output.publicpath](https://webpack.js.org/configuration/output/#outputpublicpath) to add a prefix to the relative url however it does not work with the relative import of `genesis.js` because webpack also adds the prefix there. So when `<1>_genesis.js` loads `<2>_genesis.js` will load `<base>/web_worker_<hash>/wasm_<feature>_<hash>/wasm_<feature>_<hash>/<2>_genesis.js`.

In the end, I decided to move the worker file into the respective WASM artefact folders. That works well but comes with the downside that we upload the worker multiple times (depending on the number of feature). However the size of the worker is currently at 115kb, so I don't see it as huge issue. 

[TY-2128]: https://xainag.atlassian.net/browse/TY-2128?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[TY-2127]: https://xainag.atlassian.net/browse/TY-2127?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ