TY-2283 - Improve maintainability of builder pattern
The builder pattern we are using is nice because it forbid to call build when not all the proper methods are called but it does not scale. For a new field we need to add generics in the builder for the reranker and we need to touch every methods.

[TY-2283]

Apply suggestions from code review

Co-authored-by: Robert Steiner <robertt.debug@gmail.com>

Apply suggestions from code review

Co-authored-by: Andrea Corradi <andrea.corradi@xayn.com>
Co-authored-by: Robert Steiner <robertt.debug@gmail.com>