TY-2330 add document trait (#400)
Ticket: 

- [TY-2330]

part of [TY-2282]

Summary:

Adds a `Document` trait that replaces the `Document` struct in the xayn_ai::Ranker methods. By introducing a trait we avoid the conversion of the DE `Document` struct into the xayn_ai `Document` struct. We only need to impl the trait for the DE `Document` struct.



[TY-2330]: https://xainag.atlassian.net/browse/TY-2330?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[TY-2282]: https://xainag.atlassian.net/browse/TY-2282?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ